+++ sickrage/core/databases/main/__init__.py
@@ -21,12 +21,10 @@ from __future__ import unicode_literals
 import os
 
 import sickrage
-from sickrage.core import helpers
 from sickrage.core.databases import srDatabase
 from sickrage.core.databases.main.index import MainTVShowsIndex, MainTVEpisodesIndex, MainIMDBInfoIndex, \
     MainXEMRefreshIndex, MainSceneNumberingIndex, MainIndexerMappingIndex, MainHistoryIndex, \
-    MainBlacklistIndex, MainWhitelistIndex, MainFailedSnatchHistoryIndex, MainFailedSnatchesIndex, MainVersionIndex, \
-    MainTVEpisodesByIndexerIDIndex
+    MainBlacklistIndex, MainWhitelistIndex, MainFailedSnatchHistoryIndex, MainFailedSnatchesIndex, MainVersionIndex
 
 
 class MainDB(srDatabase):
@@ -36,7 +34,6 @@ class MainDB(srDatabase):
         'version': MainVersionIndex,
         'tv_shows': MainTVShowsIndex,
         'tv_episodes': MainTVEpisodesIndex,
-        'tv_episodes_by_indexerid': MainTVEpisodesByIndexerIDIndex,
         'imdb_info': MainIMDBInfoIndex,
         'xem_refresh': MainXEMRefreshIndex,
         'scene_numbering': MainSceneNumberingIndex,
@@ -104,7 +101,6 @@ class MainDB(srDatabase):
         self.fix_dupe_shows()
         self.fix_dupe_episodes()
         self.fix_orphaned_episodes()
-        self.sync_tv_episodes_by_indexerid()
 
     def fix_show_none_types(self):
         checked = []
@@ -179,12 +175,3 @@ class MainDB(srDatabase):
             if not self.get('tv_shows', ep['showid']):
                 sickrage.app.log.info("Deleting orphan episode with id: {}".format(ep["indexerid"]))
                 self.delete(ep)
-
-    def sync_tv_episodes_by_indexerid(self):
-        show = None
-        for episode in self.all('tv_episodes'):
-            if not show or int(episode["showid"]) != show.indexerid:
-                show = helpers.findCertainShow(int(episode["showid"]))
-
-            if show:
-                show.getEpisode(int(episode['season']), int(episode['episode'])).saveToDB()